Polymorphic Posts
Well this was incredibly difficult to get working. Not only that but I
wasn't working in a separate branch so I had to make this thing work.
And the specs are green! Should be able to easily create different types
of posts all with differing functionality. I chose to use a polymorphic
association for just that reason and I suspect it will make this code
base a lot easier to work with.